+++ CHANGELOG.md
@@ -3,8 +3,11 @@
 This list is not intended to be all-encompassing - it will document major and breaking API 
 changes with their rationale when appropriate. Given version `A.B.C.D`, breaking changes are to be expected in version number increments where changes in the `A` or `B` sections:
 
+### v3.10.0.1
+- **http4k-connect-amazon-sqs-fake** - [Fix] Calculated attribute MD5 was incorrect. 
+
 ### v3.10.0.0
-- **http4k-connect-amazon-sqs** - [Breaking] Parsing of message attributes in ReceiveMessage is corrected. SQS docs are wrong... 
+- **http4k-connect-amazon-sqs** - [Breaking] Parsing of message attributes in ReceiveMessage is implemented. SQS docs are wrong... 
 
 ### v3.9.0.0
 - **http4k-connect-*** - Upgrade dependencies